Refactor login/signup component to use local state for form inputs
